YES! Finally! I did it!
I'm not sure what all is necessary to get the portlet to work, but I'm
going to drop a tag here and maybe make a zip file of all of these files
together so that I have some where to come back to. Tomorrow I can finally
start moving towards adding features and getting this thing to work.