IT WORKS!
Don't you love it when the sound track you are listening to matches up
just right with you completing a milestone in your programming? woohoo!

So it works. I've got content from either DNN or CommonSpot coming in,
take your choice. It needs some further refining, but the bones are here!
Concept PROVEN.