Progress on the more powerful DNN adapter.
I've figured out how to get a database conenction with out having to use
the USDSQL class. This is important so that I can now point that db
connection to a different location. Progress! I should be able to get this
pointing to the proper place and avoid database madness. hurray!