Expanding done!
I've got the expanding content portlet up and running. It looks
/identical/ to the current one. Much of this is because I copied the
existing code. Now it's on to tabbed content.

I also had to add a new class file. This is to provide an extra bit of
information to the portlet. I needed both page title and page content for
the expanding portlet to have all it's relevant information.

I upped the version number to 0.2. No particular reason for this. I don't
have a development road map yet. I'm going to try to come up with one next
I think. But I felt significant progress has been made on this portlet so
I upped it's value.