You can now read me!
I added a read me with a basic road map. Just trying to provide a general
idea of where to be going.

I made huge advances in the config.jsp file. No more javascript!
Everything now goes straight through the JSP and links up to the
controller on the other end. It's glorious! (Like I wish I could be on
imgur.) This brings it more in line with the way it is supposed to work.
And this all because I understand about 10% of what I'm doing now! an
order of magnitude increase since last friday. Oh yea.

I've started work on the more complicated parts of the commonspot DAO.
Because I'm having to support so many edge conditions it's not
particularly delightful to work with.