Tabs are working now. Basic JSP seperation too.
I got tabs up and working. Took me a little to get past a couple random
problems, mostly due to my copy-pasting from java to jsp. But it works
now and is pretty compacted compared to the expanding view.

I've seperated the JSP views into seperate files. I'd like the system to
be able to just pick up what files there. I'm not sure how to accomplish
that, but it would certainly be nice. It would be super cool if you
could add new .jsp views with out having to rebuild the app.