Tabs actually work now!
I spoke to soon previously. Turns out the tabs existed, but the content
was not being controlled by them. After man hours of trying all sorts of
things, I finally realized that I accidentally inserted a pound sign
into the content's ID.

I want to add the ability for content to varry depending on if it is
maximized or not. The clues I need for that can be found here:

https://github.com/Jasig/CalendarPortlet/blob/210794f3dfc2e1b761f69aeb7f353538a75e7a14/src/main/java/org/jasig/portlet/calendar/mvc/ViewSelectorDefaultImpl.java#L67