Hibernation has begun!
I've started working on hibernate in the portlet. It is coming along,
but my green-ness is showing. I keep second guessing how it should be
done. Mostly I'm falling into the trap of early optimization. Right now
I've realized that if I force pageinfo and pagecontent into one class, I
will have to request all the content for all the pages whenever I want
to get a list of pages, e.g. call getAvailablePages(). This could
present a huge performance burden.

I wonder if there is a way to not populate the entirity of a class using
hibernate? If I could just request the relevant content, that would be
grand.