It's WORKING IT'S WORKING!
Moving beyond the episode 1 quote, it's not techincally speaking working
yet. I'm making progress though. I've got a repository and all that junk
so the table just appeared in the database! I can't say as of yet that
the content is coming from the database, but it's progress! The database
is officiall hooked up.