Adding variable sources.
I've moved the configuration of the pages into an abstracted interface.
I may write an implementation of it that uses hibernate, but for now it
makes use of the portlet properties. This will result in slightly
strange looking properties, but that's ok. I've decided that this is the
best way to move forward. I can move from here into something more
complex.

I've also found a place to experiment with ajax. I want the list of
available pages to load dynamically when you change your source. I need
to learn a little about making a rest interface for portlets first.
Which I greatly approve of! Anywhere one can shoehorn in a restful
interface it is always an improvement. It could be cool to have the
content all be delievered by Ajax. That should speed up the render time
no?