My first Ajax application.
I have added a REST API to the cms content portlet. I'm amazed at how
well it is working. I've now got this portlet displaying content form
both DNN and Commonspot /side by side/. It is everything I was hoping
for! And more really. I didn't expect myself to have made a responsive
UI. Ha! Me making a half decent UI. But it works. So I'm quite pleased
with myself. I may take these principles of Ajax further with this
portlet in the future to get content so that the portlet renders faster.
We'll see, I think that won't result in an appricable speed up.

Today was a good day.