Improvments to edge cases and finishing of config.
Firstly, I have succesfully removed all references to portlet
preferences in the config and view files. Now the configuration is
completely handled by an abstract class. It currently uses the same
portlet properties, but one could do it with hibernate if so inclined.
The important thing is the complicated nature of it all is abstracted
away so the config and view classes just kinda say "Add this page" or
"remove this page" and it works with no clue on what is going on below.

Second I have dealt with a number of edge cases. What happens when there
are no pages? What happens when you add a new page and it hasn't had
it's source or path selected yet? What about when I want to shoe horn
this on to all the portlets in prod? They have only pageUri, will they
work fine?

I've also added in some cleaning steps to the propertiesDao. It will
remove excess nulls and excess channel references when ever a change is
made now. keeping things a little more stable and definately more clean.