Well it works.
I was hoping for a lot more to be done for me automatically but
noooooo. At least there isn't as much hidden mystery this way.

also, turns out the main thing I was missing yesterday was the
setter statement. Everything works now. wait... maybe there is
more to this then I realize...