Basic editing in place.
It can be a bit fritzy at times, but the basic idea works. At least
I've got content coming into the rest controller. Since I've not
got anything to write to on this branch, there isn't yet anything
that can be done with that data, but that is beside the point. I
can pretty easily move the data from the controller to somewhere
else. Besides, that is supposed to be abstracted away in the DAO.

Hopefully tomorrow I can get hibernate up and running on the other
branch so I can actually have something to write to. That would be
nice.