I've gotten it to work!
I've sucessfully pulled in data from the database through
hibernate. My biggest problem turned out to be that because I had
instantiated the internal DAO manually, the container wasn't
putting the autowirings in place. So I had to get internal DAO to
be instantiated via autowire and it all worked perfectly.

Now to begin the mergering.