Internal editor works.
I have completed the internal editor. or at very least got the
proof of concept working. You can now create and edit pages that
are stored internally in the portal from within the portal. It
does not currently have a way of deleting pages, but that can be
tollerated for now.

There is also an internal system for content and you have the
ability to choose on a per document basis what source you want the
content coming from.

Now it is time to move on to different views for maximized vs.
normal view and configuring more complex views. on to 0.5!