Different views enabled.
There isn't currently the ability to send different content, but
you can infact use different views for maximized vs minimized.
I'm pleased with how quickly this came together. My careful coding
in the previous areas certainly made it easy to extend the portlet
with this functionality. Tomorrow I work on providing different
content to the different views, and then I need to start figuring
out how to send properties to configure various features of the
views. Like placement of a document in the side bar, size of the
tabs, etc etc.