Ran into a few bumps.
I ran into some issues with hibernate. Firstly, database issue. I
wanted to save a document whose content was greater than the 8k
char limit. That means database error. I fixed this by changing the
data type to text instead of varchar(max). Now I can have something
like 2^30 or so characters. More than I'll ever need.

But changing the content was not that easy. I then had to get
hibernate to move those changes into the database. A while of
struggling with hibernate properties later I finally got it
working again. Huzzah! And it should update the prod database when
the day comes.

But before I had those issues I fiddled for a long time getting the
content to go to two seperate places. I know I am doing something
wrong because there is a lot of repeated code. Basically everything
that existed has been doubled with the addition of "max" in it's
name. Not where I want to keep this in the long run, but it helped
me get started down this path.

So now you can have different view depending on if you are
maximized or inline on the portal. For my next trick I'm going to
include the ability to select which tab is active when you open
a portlet in maximized mode. This is done through the
request.getParameterMap() method call. For example this:
cmstest.ctf3/max/render.uP?pCm=view&tab=val
gives me:
Parameter map: [tab=[Ljava.lang.String;@34ed37c7]
Fantastic! I can use this ability to provide permanent links to
tabs in portlets. This allows marketing to maintain the absolute
addressability of the content. Critical component of the portal.