Work on link copying.
Formating works properly now. I'm just struggling to get the copy
link button to work right. I think it is something that will really
make the arguement go in my favor for this idea, but clip boards
are appearently annoying to deal with. I left a bunch of code in
there commented out so I wouldn't loose it.