Copying works!
So now there is a little clip board on the tabs that allows you to
copy the direct link to that tab. I just realized I should also
do that for horizontal tabs. Anyway. this makes it easier for the
vision I had of a portlet that has different content depending on
if it is normal or maximized. Specifically, you can have the
content of the normal portlet be links to the maximized portlet on
specific tabs, or expandings, or what have you.

I think I'm going to need to rework how views work a little bit.
It is becoming clear that I'm going to need a way to associate
special information with the views. I think to do that I'll need to
make seperate classes for each view. That is at least probably the
simplest way and should allow me to hook in properties fairly
easily.

Also fixed the problem of to many options in the layout drop down.