Further work on Subscriptions
I've got the idea working for single page portlets. This is largely
due to the legacy code that still references pageUri in preferences.
I am looking to dispose of that legacy piece entirely at long last.

I'm striving to ignore the implementation details of the pref dao
as I make choices on the layout of the data and ConfigDAO. It's
working pretty well. But I've had a lot of effort racking my brains
to figure out how exactly this datamodel will work.