Broke through CMSConfig today.
I've gotten the code up to the point where I could free myself from
the burdensome code in CMSConfig. PropertiesDaoImpl droped by some
600-700 lines of code. HUGE improvement. And it's all pretty well
in line with DRY.

I'm now moving on to figure out a new interface for the config
page. I'm reading a lot about UI design as I do this hoping that I
will be able to make something actually nice. It needs a complete
overhaul. There is a lot of new features to be leveraged and a lot
of requirements that have been added since I first made that conifg
page.

I demonstrated this portlet to Ryan this monday. He was very
interested. The idea of having one less integration was enticing.
I think I've won him over to my vision on how to make this nicer.
Now to follow through on the promise of htis portlet. Make
something truly wonderful.