Progress made on new config UI.
I've got the main structure of the new config UI laid out. You will
select your view on the left, there will be a description of it and
any properties that need to be set. On the right you will have the
option of selecting the pages. I also intend on having new and edit
buttons on the page selection side which will link out to the
integrated CMSEditor portlet.

So far I have the portlet view selection working. I also have the
add button working. The document selecting for that works but does
not currently handle changing the input source. I would like to
get the change of the input source to also work with the chosen
jQuery plugin. That will allow for the content to be searched which
will help overcome the huge effort necessary to find the correct
document when there are hundreds of them.

I want to make the displaying of the subscribed documents much
nicer and more informative. Right now it just prints out the
document's toString method. I want it to list the document's title,
it's source, and it's ID. It should also have an option for
removing it from the subscriptions and for changing it's position
in the order of subscriptions, for the layouts for which it
matters.

I'm considering using tag libraries, but I'm bumping into
difficulties with importing if I do such. For example, I want to
be able to make use of c:forEach in my tags, but it is not being
particularly keen on that idea. Troublesome.