Most of the interface is done.
I made some slow progress today. But I've got most of the interface
fleshed out. Now I just need to wire up the functionality. I've got
the portlet view type updating when you change the view type to
show you the different vew descriptions, and that's about it.

Left to do: Document source changes list of documents.
List of documents is a chosen list.
Up and Down buttons for page positions.
remove button for pages.

There is the whole matter of properties, but that's down the road
a bit yet. Currently none of the views actually make use of
properties. Vertical tabs will, allowing you to set a column width
for the tab column, but it currently does not.