Progress on layout.
I've got remove and add working. The position arrows re-arrange but
I'm having trouble getting some of their related data to adjust
appropriately. Though it occurs to me that such might not even be
necessary. hmm. More to consider tomorrow.

After I get the position arrows sorted, I will need to tie them to
the backend. Right now it just changes it on the web side and the
actual properties are un-changed. So making it functional as well
as usable is the last step for those.

I think there is a problem with the view selection process. Right
now everything is ending up as view_single.