Tabs is looking good.
I've got the tabs interface refreshed. It now uses bootstrap tabs
instread of jQuery tabs. It looks a little cleaner as a result.
I've put in the copy link button there, but I haven't gone about
actually adding the code that will enable that feature. I'll tend
to that in the near future. I think I will also make that feature
optional. Have a property that you can toggle on and off. One of
the reasons I think this way is that to make the link button look
good I really have to make a button group of the tabs, which makes
it less obvious which tab is selected.

I'm now moving on to vertical tabs, but wanted to get a commit in
before I went about gutting it.