All coming together now.
Many features came online properly today. I got tabs working
yesterday, but today they really became cleaned up. Both tabs and
vertical tabs have the ability to copy a link to direct you to the
maximized form of that portlet with some permanence.

The editor got a minor upgrade. It is now possible to pass to the
editor what file you want open when you arrive. This allows the
config manager to send you to a specific document to edit. I may
also include some sort of call back so that pressing "done" returns
one to the portlet you were configuring.

Config now has it's document selector updating again. It also
updates as a chosen selector. So searching is back. New now takes
you to the cmseditor portlet. Edit does not yet work, hopefully
next monday I'll get that part of it pinned down.

The last major feature for this version is the properties. The
ability to set verious features for the different layouts. For
vertical tabs I want the ability to specifiy the width of the tabs
panel. For both vertical and normal tabs I want the ability to
toggle the copy link button on and off.

Also, worth remembering: I need to get maximized configuration back
into this. ^^;