The dawn of 0.5!
I'm calling an end to the 0.4.x line of development. The
configuration side of the portlet is now fully operational, and
looking better than ever. It's was smoother and simpler to use and
doesn't have any full page reloads for any of the actions.

You can open the editor on a new document or on a document that
currently exists. You can click add and remove on any document to
easily add or remove. Arrows now provide a simple system for
changing the order of the documents. The new config UI is great!

Tomorrow I will pick it over with Ryan as we consider building
interest behind this portlet with the upper ups. I hope that my
effort here will make this a very convincing message.

Moving forward, I have a few more layouts I'd like to add for
verity. It would be good if you could read in documents from one
source in to the editor and save it to another. Particularly if
you could pull it in from a non-save capable resource (CS) and then
save it to a save-capable resource (internal). One of the main
desires of the major refactoring I did with layouts and
subscriptions was to add properties. That will hopefully be
something I can work out in the next version.

Also hopefully I'll have security, per-page security, coming with
the next version. Layouts and subscriptions has made it a
possibility.