Progress on properties.
It took some mental effort, but I've made progress on properties.
Figuring out how exactly to bend java to my will was the challenge.
But in the end I figured out how to get a new object of the correct
type and have it's methods be called. This allows for default
properties to be set in the layout itself, keeping that abstract
so that the base class doesn't have to worry about it.