Properties in place. Ryans requests are fulfilled.
Ryan wanted me to make a couple changes before we show case what I
have made thus far. I've completed those changes now and I think
it is ready to show.

Specifically I've added the ability to choose whether tabs in
vertical tabs has links or not. I've made it so that if you click
to edit a document in the internal editor, it will direct you back
when you are done.