Done! Security now works.
I just had a get the last little bits today, which took a little
longer than hoped. But it works now.

Based off of current winds, I have decided to abandon putitng in
extra layouts for now. I may have to throw some back in to what
we need. Also, forms is something new that will be needed to be
added to the system.