Generic responders.
I've setup a system for writing generic responders to forms. I'm
pleased with how my understanding of beans has really progressed.
It was fairly straight forward to achieve what I wanted. Also today
I added the ability to edit previous forms to the form editor. So
now it is an editor not a creator. There are still some features
that need to be added, like removing rows and re-arranging. But it
is coming along nicely. Data now flows into the server and is sent
to a generic processor. Tomorrow I will work on getting that to
send an email. The database is just a holder for now. I'm not sure
if it will actually be used.