Edit/new buttons in place
I've got the new buttons in place and wired correctly. I also
updated a little bit of the code that backs it. When you click on
"edit" it will direct you to the correct editor.

I'm trying to fix the other issues with the editor now to complete
this alpha. Namely adding the ability to delete and fixing the
problem where you had to save twice in order for the document to
actually save. The later appears to be because I'm using normal
form rather than AJAX like I did for the form editor. No matter, I
will fix that tomorrow and then it will be working correctly.

The delete issue though is being problematic. I don't think I'll
get it done this alpha. The problem is somewhere deep in the way
hibernate communicates with MSSQL. I'll have to dig in to really
figure out what's going on there exactly.