Progress on search.
I've progressed a little on the search front. I need to shift gears
for a moment and work on some other things, but I'll be back.

In this commit I modified the portlet.xml and CMSContent.xml files
to try to get the portal to realize I have portlet events for
search in this portlet. I also setup the portlet to randomly return
results. This is just a place holder till I get actual search in
place. But I've yet to get the results to actually return so that
is the important thing right now.