Doc Type cometh!
I'm creating a proper system for determining document type. It will
be stored in the database for the internal documents, but for
other document sources you are expected to just provide the
document type in the DAO. Because most DAO will only provide one
document type, that shouldn't be very complicated. I hope to use
this system later to allow for videos to be integrated into the
system. That would be snazy.