Content retrieval stream lined.
Sort of at least. I moved the retrieval of content from the
controller into the CMSLayout class. This is done in preparation
for merging with the search branch. The code that was used for
retreiving was reproduced in both of the controllers. So moving it
to the central place of the CMSLayout class makes the code a little
more DRY.

Last thing to do before merging is to comment that process. It is a
long convoluted process that is not easy to understand. It would be
nice if I could also break it down into some methods. Right now it
has 3 nested for loops at it's worst.