Cleaned up code nicely.
That process turned out to be easy to clean up because of the move
of the JSON stuff into the CMSDocument class. So since the doc now
handles that I could just cut it out, and that was probably half of
the getContent process. But I trimed it down a little further and
now there is just two for loops and both make sense.