Well that was simple.
I went to update the search controller to use the new mechanism
for getting content and it was replacing many lines with one. w00t!
I think I'm ready to merge this into rewrite and merge rewrite into
master. I'm terminating the rewrite branch. This is no longer a
rewrite this is the portlet. In the future branches will be named
for the feature they add or change. I may still make a "patches"
branch, but the point of version controll is I can choose which
version to use in production so patch branches shouldn't be
necessary either.