Version 0.7 complete.
I've gotten all the requirements for version 0.7 completed. Time to
move on to version 0.8. Almost done with this project. I can't
believe what a vast piece of work it has become. I really hope that
it serves USD well for years to come.

Ironically, the next step in this process is not actually part of
this portet so much. My next step is to go and make a GUI for the
attachments portlet so that I can then use that portlet to host
files in the portal and this portlet can attach to that service.