Added CS Portal pages document source.
I made a new source out of the existing portal pages. CSPortalPage.
This will allow marketing to tap into the existing portal pages
much easier and allow them to be transfered easier then.

There was also an issue with the rest api that was caused by
CMSDocument having a method starting with "get". I changed
"getContentJson" to "json" and that solved the problem.