Rough waters.
I've run aground. SOmething is up with the documents, I'm getting
a lot of null pointer errors. At first the problem was that I had
dropped and re-created the database and forgot to copy the docs
over to dev. Fixed that issue and the problem seemed to go away,
but now it's back. I don't have enough time to really dig into this
today. I'll work on it more tomorrow.