Editor tweaks underway.
I've tweaked the editor a bit to make it more useful. But first I
should mention that I fixed a problem with the config scrren. It
wouldn't show documents from whatever the source was when you
loaded the page. It would show internal regardless. This leads to
an awkward situation where you could add a document that only
existed in internal but with the CSPortalPage data source.

That fixed I moved on to the editor. One of the big features I
wanted for the migration process was the ability to open a document
in one data source and then save it in another. Now, of course,
only one data source manages saving documents, the others are read
only, but that will work for our purpose. This should permit
marketing to open a document from CSPortalPage and then save it to
internal. That should save some effort.

Now I want to re-do the UI for the editor. I want it to have a
proper file browser so that marketing gets the id/path correct
when they are making their files. But I figured I'd commit my code
before I went and screwed it up.