Editor now searches.
Got search working today. It is a fuzzy auto search, so just start
typing and it will start searching. I've been working a lot on the
tree now. I decided that click and load is not the best way to go
about loading documents. Instead they will click on a document and
then click the load button. this also allows for documents to be
saved over other documents.