I need to re-think my approach
I've been trying to setup a system for selecting a file in the tree
when you load from the configuration manager. It has not been going
well. But I just had an idea that might help me get around this
problem. If I pass the path to each node, I should be able to
request the node by that data and make it active. That should allow
the correct folder to be selected when you load.

I've tried doing it through a number of other means in the
javascript but they've all failed for technical reasons. Things
like that part of the document isn't loaded yet.