Per-Document search added.
I've added the ability to insert search terms into documents. This
is a new data member in the document that can be optionally filled
by a DAO. Currently of course, only internal uses it. I don't see
a reason at this point to enable it for the others. If you choose
to not enable it on a DAO, that makes very little difference. The
search results for your DAO just won't benefit from the keyword
matching.