Forms a fancy!
I've got forms functioning correctly now. You can open an edit them
and save them and they respond correctly. I also added the ability
to have responders have some minimal level of configurability. It
is so far mostly useful for the email response type. You can enter
the email address that is sending the email as well as the email
address that will be recieving.