Nearing complete fully featured portlet.
There is spaghetti in there, it could be cleaner, but the code base
isn't the worst I've seen. I'm closing in on covering all of the
things. ALL THE THINGS! :D

Today I fixed the issue with layouts not updating. Turns out it was
just a problem with Vertical Tabs. When the copy process happens it
needs to not create a new instance of itself, but rather a new
instance of CMSLayout. This may still not be quite right. I'll have
to fiddle a little more with how exactly this whole inheritance
process works.

Fixed the issue where new documents were being added in the config
screen without security groups.

This is ready now for production tomorrow. With this push up the
process of migrating content from CommonSpot to CMSContent's
internal begins!