Starting Lucene integration.
I've got Lucene up and searching, but I need to get the handoff on
the other end more tightly integrated. Right now every portlet
finds a result and returns it as if I matched itself.

There are a couple of performance gains I think that could be made
as well. And lastly the keyword matching needs to highlight docs.