Fixed properties and new layout.
I added a layout that looks like the old CMSContentTabbed, because
it largely is, so that the transition from CMSContentTabbed to
CMSContent can go a little more smoothly.

I also fixed the problem with properties. Tabbed didn't have the
logic in place to have it's properties do anything, so fixed that,
also fixed default properties not showing up.