Internal rendering is now a thing!
I've spent all of yesterday and so far today working on this one
problem: Providing a mechanism to render JSP's to a string so that
I can use JSPs for document sources that need to be rendered before
they get passed in as content on a document.

This is specifically important to what I am doing with forms. My
objective was to turn forms into their own document source to keep
things modular. No longer will the display process care about what
type of content a document is, that will be the document's job.