Form enhancements falling a little short.
I have a completely different system for form enhancements on a
different branch, but I think they were a little overboard. I've
pruned it back quite a bit and have something nearing completion
now. This version of my form enhancements moves forms to their own
document source. I wanted to merge the editor, and I still may, but
right now it was proving to be more trouble than it's worth.

The important thing is I've got forms full operational again with
the new document source. The last thing I want to do on this
particular branch is make the database responder.