Form responses now working.
I've spent a while trying to get hibernate to work. I'm not
entirely sure at what point I did it right and how it was different
from what I had done previously. But it works now. Two new tables,
one for responses and the other for the answers for the response.