Fixed a couple minor errors.
The edit button wasn't directing forms to the form editor. This was
due to the contentless query assuming everything was html. Fixed
that and now the edit button works as expected.

Title was not saving with forms. Some things just weren't hookedup.

Similarly I fixed the load form on load feature of the form editor
which wasn't working. Again, mostly some things that weren't
hooked up or at least not hooked up correctly.