Fixed multiple documents issue.
Was exclusive I believe to CSPortalPage. Fixed the problem by
adding another column to the view where the documents are sourced
from. I then sort on that view by the newly added document approval
date and then select only the most recent. This reduced the number
of documents from 500 to 300ish. Nice.