New folder fixed.
New folder wasn't working because of a tweak I had made in the way
information is saved when nodes are created. The new folder didn't
get that tweak so it wasn't working properly. After adding this new
bit of code to the function it works fine.

I also updated the save function to display new documents upon
saving them. And lastly search now hides all the nodes that don't
match what you are searching for.