Email working.
Turns out the only thing really wrong with the email responder was
that it was pointing to the wrong address. It's now stmp.usd.edu.
Once that was in place it worked. Though it takes a really really
long time to submit, 50-90 seconds. So I also upgraded the submit
button. It doesn't just disable now, it disappears. And is replaced
by a loading gif.