Started work on componentization at the same time.
I've decided to work on componentization along side getting reports
up and running. I'm doing this largely because of the editors. I
want there to be only one editor but with three editor types. So
when you change your data source to a different document type, be
it form, html, or report, it updates appropriately. This sounds
like a big old headache with the current way things are setup. But
if I can componentize it all, then it shouldn't be to hard to do
and will make my program all the more sleek.

Also setup basic stuff for the reports data source.