Vertical tabs a fancy!
I've made a few upgrades to the vertical tabs layout. Now when you
select a tab it updates your address bar so that you can link to
the tab you currently have open. This is much better than the
copy to clipboard feature since it doesn't add an extra button and
isn't dependant on finicky libraries.

I also fixed a small problem in the get contentless methods of the
internal data store.