New API.
I started to work on breaking up doc-tree by making doc-selector,
which lead me down the rabbit hole of rebuilding the API. So now
there is a smaller cleaner api that actually is more functional. It
will require web components to work entirely as planned, but that
is what I'm going for so that should be fine.

I've got the basics of the doc-selector working, but right now I'm
trying to get the on source change portion of it working. It has so
far remained entirely a mystery. I hope to get this solved
tomorrow.