Public version time.
I'm finally confident that I can publish this to github without
revealing sensative USD information and it may even be of some use
to other universities. The main effort was getting this portlet to
a position where it could be overlaid properly. I also updated the
readme extensively this commit to provide information to anyone
who happens across the portlet. I intend to push this to the
uportal contrib group, the readme should help anyone looking to
adopt to identify if it's something they are interested in.