Realizations sinking in.
I'm having some challenges with some of the things I'm trying to do
to get webcomponents to work. I talked with Christian Murphy about
it a little on slack and the biggest realization I discovered is
that if I wait to do this till USD is on uP5, then I can leverage
NPM for getting access to webcomponents from within my web
components. But I can use Raw git for the time being. The biggest
realization I've hit on though is that of the seperation of
reports from this process. I figured out a way to do reports with
out having to have a reports editor, so I'm going to go down that
road for the time being. This will allow me to close the common
spot replacement project which is my highest priority right now.

But hey, I got doc-selector to talk to doc-tree!