Started work on Reports.
I'm going to have reports work by simply providing a table of
results based off of what forms exist. It's not the ideal
approach, but it is short and sweet and to the point. Once I figure
out how to do the web components thing correctly, I'll probably
make reports a little more fancy, but for now it provides a
solution to the inability to see reports.