New responder type and multiple resonders.
I finally got around to enabling multiple responders. This turned
out to be the simple way to also include a confirmation email. I
just made a new responder type of "Confirmation Email" and all was
well. But! Now though I think there is a little more that needs to
be done. I think I should store all responses in the database and
then have responders react to them after that. The reason being is
that some responders can take a while. And if we are emailing 2-3
people it could be several minutes before the user gets the feed
back that his form was recieved. So If I save it in the database,
the response is saved. Then from the database, we can process it
out into the other responders on a seperate thread.

That will be a bit of effort so it will probably take a bit before
I actually go down that road.