Fixed vertical tab content placement.
I also removed jquery and bootstrap from the mix in the process.
It is overall a cleaner code base now. I've needed to rewrite this
layout for a while so I'm glad for the opporunity. This was brought
to the forefront by an issue where the content was not being placed
properly in the layout.