Multithreaded responding now in place.
I have wired up the form api to do it's responding to form subs in
a multithreaded way. The reason for this is to allow responders
that are slow to respond to not block the UI. Also I introduced the
concept of "autoResponder" which determines if a responder is to
respond before returning to the user or after. Basically the only
thing that should be an autoResponder is the database. But I wanted
to do such in an abstract way so I wouldn't be tying directly to a
specific responder.

Also some tweaks on the expanding layout to make it's buttons red
as long as the section is open.