Datetime picker in place.
I couldn't for the life of me get the silly jQuery plugin to work.
So I went the route you never should and rolled my own. It's not
amazing but boot strap makes it look pretty ok. It meets all the
requirements while still functioning just fine. So I'm calling it
good enough.