Removal error fixed.
When a new subscription was made, the row for that subscription in
the UI wasn't being properly filled. A minor tweak was all that was
needed to get the relevant information into place. A edit I must
have over looked previously.

But before I fixed the error I fixed the greater problem. If an
incomplete request was sent to the portal for a document being
added or removed, it would cause problems. Specifically it was
resulting in the whole layout being deleted and problems cropping
up. So I put a check in there so that the explosion won't happen
again even if I haven't fully fixed the UI problem mentioned above.