Improved email responder to hanlde errors.
There was a form that was failing everytime it was submitted on
just the email responder. While I was at it I updated the table
display a little. They didn't want the field to be truncated so I
removed that feature. And then I also added cell boards because it
became hard to read with the full field names there.