Database saving enabled.
I don't usually go this long without making a commit, it's rather
odd of me. So in this one commit a lot happened. Several files were
changed in non-trivial ways to allow for content to be saved
directly to the database rather than with portlet properties. This
is still very much so an experimental idea, but in doing it allows
for some features that were not viable previously. Most notably is
the ability to edit the portlet's layout without having to go into
configration mode.

My hope is to be able to make a restful api the manages the actual
configuration of the portlet. There are some security concerns with
that idea, but we'll see what happens. I'm not sure I actually like
this approach to saving the content for the portlet. I don't think
it is any less prone to bugs than the properties appraoch, which is
something I had hoped it would be. Hibernate is nice and powerful
but it has to work it's way. This also adds CMSProperties,
CMSLayout, CMSSubscription, and CMSSecurityGroups tables to the
database.

The idea is that instead of using the properties mechanism that is
built into the portal, everything is a hibernate object. And
because of that, all you would need to be able to safely edit the
layout is the fname of the portlet that displays it. This is the
key difference. With the properties approach you can really only
edit it if you are in config mode on the portlet itself. This is
very had to wire up with a rest API. Also, SpringMVC is going away
so all of the bindings I'm using will also be disappearing. This
approach future proofs against that.

There is a lot of things going on here, and my mind is spinning
quite a bit. I'm not sure what the appropriate next step is and so
I am just experimenting with the options.